+++ x/observer/keeper/voting.go
@@ -46,27 +46,26 @@ func (k Keeper) CheckIfFinalizingVote(ctx sdk.Context, ballot types.Ballot) (typ
 func (k Keeper) IsValidObserver(ctx sdk.Context, address string) error {
 	isActiveObserver := k.IsAddressPartOfObserverSet(ctx, address)
 	if !isActiveObserver {
-		return sdkerrors.Wrapf(types.ErrNotObserver, " Observer address is not part of the observer set: %s", address)
+		return sdkerrors.Wrapf(types.ErrNotObserver, "address is not part of the observer set: %s", address)
 	}
 	valAddress, err := types.GetOperatorAddressFromAccAddress(address)
 	if err != nil {
-		return sdkerrors.Wrapf(types.ErrInvalidAddress, "Invalid operator address: %s", address)
+		return sdkerrors.Wrapf(types.ErrInvalidAddress, "invalid operator address for observer : %s", address)
 	}
 	validator, err := k.stakingKeeper.GetValidator(ctx, valAddress)
 	if err != nil {
-		return sdkerrors.Wrapf(types.ErrNotValidator, "Validator not found for observer address: %s", address)
+		return sdkerrors.Wrapf(types.ErrNotValidator, "observer is not a validator: %s", address)
+	}
+	if validator.Jailed {
+		return sdkerrors.Wrapf(types.ErrValidatorJailed, "observer is jailed: %s", address)
 	}
 	consAddress, err := validator.GetConsAddr()
 	if err != nil {
-		return sdkerrors.Wrapf(types.ErrInvalidAddress, "Invalid consensus address for validator: %s", address)
-	}
-	if validator.Jailed {
-		return sdkerrors.Wrapf(types.ErrValidatorJailed, "Observer address is jailed: %s", address)
+		return sdkerrors.Wrapf(types.ErrInvalidAddress, "invalid consensus address for observer: %s", address)
 	}
 	if k.slashingKeeper.IsTombstoned(ctx, consAddress) {
-		return sdkerrors.Wrapf(types.ErrValidatorTombstoned, "Observer address is tombstoned: %s", address)
+		return sdkerrors.Wrapf(types.ErrValidatorTombstoned, "observer is tombstoned: %s", address)
 	}
-
 	return nil
 }